Add precondition support to DELETE requests

+++ java/org/apache/catalina/servlets/DefaultServlet.java
@@ -703,6 +703,10 @@ public class DefaultServlet extends HttpServlet {
 
         WebResource resource = resources.getResource(path);
 
+        if (!checkIfHeaders(req, resp, resource)) {
+            return;
+        }
+
         if (resource.exists()) {
             if (resource.delete()) {
                 resp.setStatus(HttpServletResponse.SC_NO_CONTENT);
@@ -712,7 +716,6 @@ public class DefaultServlet extends HttpServlet {
         } else {
             resp.sendError(HttpServletResponse.SC_NOT_FOUND);
         }
-
     }
